ESPiLight::send: use size_t for repeats

+++ src/ESPiLight.cpp
@@ -201,7 +201,8 @@ void ESPiLight::sendPulseTrain(const uint16_t *pulses, size_t length,
   }
 }
 
-int ESPiLight::send(const String &protocol, const String &json, int repeats) {
+int ESPiLight::send(const String &protocol, const String &json,
+                    size_t repeats) {
   if (_outputPin < 0) {
     DebugLn("No output pin set, cannot send");
     return -1;
